Update src/solver/impls/snes/snes.cxx
Code changes suggested by ZedThree

+++ src/solver/impls/snes/snes.cxx
@@ -142,7 +142,7 @@ SNESSolver::SNESSolver(Options* opts)
           (*options)["pidController"].doc("Use PID controller?").withDefault(false)),
       target_its((*options)["target_its"]
                      .doc("Target snes iterations")
-                     .withDefault(static_cast<int>(7))),
+                     .withDefault(7)),
       kP((*options)["kP"].doc("Proportional PID parameter").withDefault(0.7)),
       kI((*options)["kI"].doc("Integral PID parameter").withDefault(0.3)),
       kD((*options)["kD"].doc("Derivative PID parameter").withDefault(0.2)),
